requirements.txt
biopython>=1.68
numpy>=1.12.0
matplotlib>=2.0.0
Sphinx>=1.5.1
